add authorize for policy to category module

+++ Modules/Category/Http/Controllers/Panel/CategoryController.php
@@ -2,8 +2,11 @@
 
 namespace Modules\Category\Http\Controllers\Panel;
 
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\RedirectResponse;
 use Modules\Category\Enums\CategoryStatusEnum;
 use Modules\Category\Http\Requests\CategoryRequest;
+use Modules\Category\Models\Category;
 use Modules\Category\Repositories\CategoryRepo;
 use Modules\Category\Services\CategoryService;
 use Modules\Share\Http\Controllers\Controller;
@@ -12,6 +15,8 @@ use Modules\Share\Services\ShareService;
 
 class CategoryController extends Controller
 {
+    private string $class = Category::class;
+
     public CategoryService $service;
     public CategoryRepo $repo;
 
@@ -25,9 +30,11 @@ class CategoryController extends Controller
      * Read data with show list of categories.
      *
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('manage', $this->class);
         $categories = $this->repo->index()->paginate();
 
         return view('Category::Panel.index', compact('categories'));
@@ -37,9 +44,11 @@ class CategoryController extends Controller
      * Show create category page.
      *
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize('manage', $this->class);
         $parents = $this->repo->index()->get();
 
         return view('Category::Panel.create', compact('parents'));
@@ -47,9 +56,14 @@ class CategoryController extends Controller
 
     /**
      * Store category with show message & redirect.
+     *
+     * @param  CategoryRequest $request
+     * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(CategoryRequest $request)
     {
+        $this->authorize('manage', $this->class);
         $this->service->store($request);
 
         return $this->successMessageWithRedirect('Create category');
@@ -60,9 +74,11 @@ class CategoryController extends Controller
      *
      * @param  $id
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @throws AuthorizationException
      */
     public function edit($id)
     {
+        $this->authorize('manage', $this->class);
         $category = $this->repo->findById($id);
         $parents = $this->repo->index()->where('id', '!=', $category->id)->get();
 
@@ -72,12 +88,14 @@ class CategoryController extends Controller
     /**
      * Update category by id.
      *
-     * @param  CategoryRequest $request
+     * @param CategoryRequest $request
      * @param  $id
      * @return \Illuminate\Http\RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(CategoryRequest $request, $id)
     {
+        $this->authorize('manage', $this->class);
         $this->service->update($request, $id);
 
         return $this->successMessageWithRedirect('Update category');
@@ -88,9 +106,11 @@ class CategoryController extends Controller
      *
      * @param  $id
      * @return \Illuminate\Http\JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->delete($id);
 
         return AjaxResponses::SuccessResponse();
@@ -101,9 +121,11 @@ class CategoryController extends Controller
      *
      * @param  $id
      * @return \Illuminate\Http\JsonResponse
+     * @throws AuthorizationException
      */
     public function active($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->changeStatus($id, CategoryStatusEnum::STATUS_ACTIVE->value);
 
         return AjaxResponses::SuccessResponse();
@@ -114,9 +136,11 @@ class CategoryController extends Controller
      *
      * @param  $id
      * @return \Illuminate\Http\JsonResponse
+     * @throws AuthorizationException
      */
     public function inactive($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->changeStatus($id, CategoryStatusEnum::STATUS_INACTIVE->value);
 
         return AjaxResponses::SuccessResponse();
